requirements.txt
langchain-openai>=0.0.5
langchain-pinecone>=0.0.1
pinecone-client>=2.2.4
python-dotenv>=1.0.0
google-api-python-client>=2.108.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.2.0
langgraph>=0.0.20
pydantic>=2.0.0
langsmith>=0.0.50 